fix: hide impl details, split modules, improve code for analysis
